Add very simple import blobs script
I know there is already one but I just made this to get some libs in
case I am too lazy to use the other one. (How was I not lazy enough not
to make this script? Wtf)
Note that it does not deodex yet, so don't import any apks or jars
using this.

Change-Id: Ic2ab9e51da1af9b80308ee20d7e66d3f568bea74